feat(web): refactor files
