+++ packages/web/src/new/scenes/Main/Buy/Subviews/Moonpay/MoonpayIframeWidget.tsx
@@ -6,10 +6,10 @@ import { observer } from 'mobx-react-lite';
 
 import { useTrackEventOpen } from 'app/hooks/metrics';
 import type { BuyViewModelProps } from 'new/scenes/Main/Buy/Subviews/Moonpay/types';
+import { WidgetPageBuy } from 'new/scenes/Main/Buy/Subviews/Moonpay/WidgetPageBuy';
 import { MOONPAY_API_KEY, MOONPAY_SIGNER_URL } from 'new/services/BuyService/constants';
 import { buildParams } from 'new/services/BuyService/MoonpayProvider/utils';
 import type { MoonpayIframeParams } from 'new/services/BuyService/types';
-import { WidgetPageBuy } from 'new/ui/components/pages/buy/WidgetPageBuy';
 
 const Wrapper = styled.div`
   height: 640px;
